+++ src/index.ts
@@ -8,6 +8,19 @@ if (!process[Symbol.for('ts-node.register.instance')]) {
 
 // Stack traces
 Error.stackTraceLimit = Infinity;
+const nativePrepareStackTrace = Error.prepareStackTrace;
+Error.prepareStackTrace = (err, traces) => {
+	return nativePrepareStackTrace(err, traces)
+		.split('\n')
+		.filter(line => {
+			return (
+				line &&
+				!line.includes('node_modules') &&
+				!line.includes('internal/process/')
+			);
+		})
+		.join('\n');
+};
 
 // Enter application
 import './main';