+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 - [TSTEMPLATE-7] Source mapping references wrong line numbers when run in dev/ts-node
 - [TSTEMPLATE-8] Set stack trace limit to Infinity
+- [TSTEMPLATE-9] Filter stack traces for "just my code"
 
 ## [v1.2.1]